voice_chat/requirements.txt
streamlit==1.32.2
streamlit-audiorec==0.1.3
transformers==4.39.0
pydub==0.25.1
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1+cu118
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
faster-whisper==1.0.1
datasets==2.18.0
ctranslate2==4.1.0